% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map State Accident}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{years}{The year of concern}
}
\value{
\code{fars_map_state} will return a map plot of accidents for the given state and year.
  If an invalid state number is provided an error will be returned and if no accidents accurred
  in that state for that year a notification will be provided.
}
\description{
\code{fars_map_state} will plot the accidents on a map for a given state
  and year.
}
\examples{
fars_map_state(2, 2014)

}
\seealso{
\code{\link{make_filename}} to understand how the file name is created

\code{\link{fars_read}} to understand how the the file ir read in
}
