% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Observation by Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The yaers relating to the file namesto be read in}
}
\value{
\code{fars_summarize_yaers} will returned a wide formated data frame.

\code{\link{fars_read_yaers}} to understand how the file name is created.
}
\description{
\code{fars_summarize_years} will read in multiple Fatality Analysis Reporting
  System data files based on the years provided and summarize the number of
  observations by month and year.
}
\examples{
fars_summarize_years(2015:2016)

}
